Move `transcribestreaming` test utility to `aws-smithy-eventstream`
