+++ rust-runtime/aws-smithy-eventstream/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "aws-smithy-eventstream"
 # <IMPORTANT> Only patch releases can be made to this runtime crate until https://github.com/smithy-lang/smithy-rs/issues/3370 is resolved
-version = "0.60.6"
+version = "0.60.7"
 # </IMPORTANT>
 authors = ["AWS Rust SDK Team <aws-sdk-rust@amazon.com>", "John DiSanti <jdisanti@amazon.com>"]
 description = "Event stream logic for smithy-rs."
@@ -11,6 +11,7 @@ repository = "https://github.com/smithy-lang/smithy-rs"
 
 [features]
 derive-arbitrary = ["arbitrary", "derive_arbitrary"]
+test-util = []
 
 [dependencies]
 arbitrary = { version = "1.3", optional = true }
